[RENAME] JwtTokenProvider 클래스명 JwtProvider로 변경
Jwt(JSON Web Tokens)와 Token의 의미가 중복되어 클래스명 변경
